Work out 'unable to resolve inferred error set' error during compilation
The solution was simply to use an explicit error definition with all the
possible error values that could come up from the entire call stack,
even in recursive fashion. That basically made the compiler happy.
(The language doc says this could change in the future, since the
explicit error definition is kind of a workaround for the time being)