Fix a bug when printing the Environment hashmap.
Kinda caught me off guard, but shouldn't be surprising in the end.
It's just a simple case of literal string value ownership.
The variable statements (which the Environment uses to define the scope
variable entries) are using the literal string pointers from the Tokens'
lexme returned by the Parser.
In other words, the lifetime of those lexeme values are all jumbled up,
and thus the Environment debug printing seems to be crashing for no
apparent reason.
All I had to do was "clone" the lexeme from the token into the variable
statement, effectively making it owned by the statement itself.