A few code reorganizations and clean ups.
The Parser and Interpreter types are now refactored as module structs,
which in turn allows their many functions to sit nicely at the top level
of the file.
Zig's modules (files) double-served as implicit types, so at the end of
the day, a type is a type and can be instantiated. There is no semantic
difference with explicitly giving it a name and defining it.