Got the `LoxCallable` type composition working,
after a bit of finagling.
But, it cannot distinguish between built-in provided (function pointer)
calls, to the actual function definitions in the scripts (function
statements). Function bodies are being executed no matter what.
Gonna have to work on that!

In summary, the built-in functions (such as `print` and `clock`) will
use the function pointer to compose the LoxCallabe, while the script
function definitions will be defined as part of the current
`Environment` scope and executed.