A big refactor on token `consume()` with error reporting upon fail.
Factor out the user error reporting to a new module on its own, so that
it can be called from anywhere.
The token consume-with-error-report mechanism is a lot cleaner then
before now.
Refactoring it kinda fixes some (only "some") of the wonkyness of
parsing function statements. At least, the function name identifier is
recognized properly now.
Some scoping issues on the argument variable bindings remains. But
that's good, since I don't expect the function definitions and calling
to be functional (pun-intended) yet.

Clean up and refactor the `synchronize()` routines too, which fixes some
of the weird bugs.

Overall, not much big progresses, but feels good.