Fix the NEX-1719 bug, the declared function *disappeared* from env
Revert the arena allocator initialization in main back to GPA for the
time being. There will be a separate refactoring pass of cleaning out
and tidying up the allocation spaces down the line, but now's not the
day yet.

Add a small control to show/not show the "quotes" around the string
values. The interpreter outputs of the string values doesn't need quotes
obviously, but the debug outputs and envrionment outputs are better
served with quotes around the strings.
Clean up a lot of allocated `toString()` function calls just to debug
display the output of certain things, like Values and EnvValues.

More details about the bug fix will be in the PR notes.